refactor (countdown timer): separate non-react JS into lib
